[DEV] Adjust "Character MoveSpeed Logic" // DEV_86
[DEV_86] 캐릭터 MoveSpeed 로직 변경에 따른, 무기Database의 M Weight, WeaponBaseClass 수정

+++ NewAWWA/AWWA/Source/AWWA/Public/BaseWeapon.h
@@ -60,7 +60,7 @@ public:
 	// Init in Constructor! from Database! to reduce Speed
 
 	UPROPERTY(BlueprintReadWrite, EditDefaultsOnly, Category = "Stat")
-	int32 m_weight;
+	float m_weight;
 	// Init in Constructor! from Database! to reduce Speed
 
 	UPROPERTY(BlueprintReadWrite, EditDefaultsOnly, Category = "Stat")
